Copy the Upload ID to the clipboard
- Add a button to copy the Upload ID to the clipboard.  Message is integrated,
so not very specific as before. Maybe submission process will be revised, so I'd
hope this is okay.

- In addition, I modify the prefix of `astropy.table.Table` object as `tb_`
in some places in `pn_app.py`. Maybe this is imcomplete, but keep this as
the future refactoring work as it is really confused when `df_` is not used
for objects not `pandas.DataFrame`.