Reset FileInput instance when a session starts
Moritani-san detected a strange behavior that targets of a previous
session appeared to be displayed in the list even when no file is
selected in the FileInput widget.

It seems that the values are kept between sessions because the
FileInput class does not have a constructor.

I added a reset function rather than adding a constructor because
the watcher function for the update of the value does not seem to
work when the file_input instance is created. The initial
token is not generated as well.

I'm wondering if there are other ways to implement this with a
constructor.

The issue that non-Gecko rendering engines does not work with the
watcher function. I have no idea what to do at this moment. I'm
asking questions in the Panel forum.