fix: Force to sleep for a shoft time for Tabulator
Sometimes the target list table fails to load data content after the
validation. It looks that the error is raised when some operations
(e.g., pressing the validation button again) are commanded before updating
the DataFrame in the Tabulator widget for targets.

This commit forces to sleep for a short period (0.1 and 0.2 s) for these
processes. I hope it fixes the issue.