Replace hvplot.scatter to hv.Scatter in a binded function
It seems that hvplot.scatter() is much slower than holoviews.Scatter(). I tried to replace the former to the latter in the callback function in plot_ppc. From the profiling, the rendering appears to be faster. Still, response is not super fast. I'm not sure if we can make "pipelined_call" to be faster.