Configurable exposure time for individual exposure
Add a checkbox to indicate simulating a classical proposal and
integer input form to specify exposure time per visit. It is in the
range between 10s to 7200s with the default to 900s.

The number is propagated down to the uploaded files as a meta data.

Documentation update for PPP with a custom single exptime
- Add description on the optional configuration for the individual exptime.
- Add a list of some constant parameters used in the simulation for transparency to the community.