Possible fix for `BrokenPipeError` when `max_exetime` is reached
It seems that the multiprocessing version of `KDE()` calculation
is not terminated properly before terminating the PPP calculation
which also uses multiprocessing module. Apparently, nesting multiprocessing
would cause the issue when the parent process is killed before all
child processes finishes.

This seems fix the issue, though the issue appears only for macOS.